backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "psutil>=7.2.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pyobjc-framework-metal>=12.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]
